Write a function that prints half of a string, followed by a new line.

+++ 0x05-pointers_arrays_strings/7-puts_half.c
@@ -9,7 +9,7 @@
 
 void puts_half(char *str)
 {
-	int len = strlen(str), i;
+	long int len = strlen(str), i;
 
 	if (len % 2 != 0)
 		len += 1;
